added tool for creating mng format

+++ src/System/RenderSystem/d3d9IndexData.cpp
@@ -21,8 +21,8 @@ void	D3D9IndexData::lock(unsigned short* &_indecies, size_t _offset, size_t _cou
 	m_indexCount = _count;
 	m_indexStart = _offset * 3;
 	m_IB->Lock(
-		m_indexStart * sizeof(unsigned short), 
-		m_indexCount * sizeof(unsigned short), 
+		(UINT)(m_indexStart * sizeof(unsigned short)), 
+		(UINT)(m_indexCount * sizeof(unsigned short)), 
 		(void**)&_indecies, 
 		D3DLOCK_NOOVERWRITE);
 }
@@ -40,7 +40,7 @@ void	D3D9IndexData::createIndexBuffer(size_t _count, size_t _start)
 	HRESULT hr;
 
 	hr = m_pd3dDevice->CreateIndexBuffer(
-		m_indexCount * sizeof(WORD), 
+		(UINT)(m_indexCount * sizeof(WORD)), 
 		0, 
 		D3DFMT_INDEX16,
 		D3DPOOL_MANAGED,
